backend/static/ezhealth_final/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
bcrypt==4.1.2
Pillow
torch
torchvision
pydantic[email]
numpy
opencv-python
matplotlib
joblib
pandas
scikit-learn==1.3.2
